requirements.txt
beautifulsoup4
hikari
hikari_lightbulb
pandas
requests
tqdm
pydantic-settings